Reorganize benchmarks file
